# backend/openapi.yaml
openapi: 3.0.3
info:
  title: FinTrak
  description: |
    Personal finance tracker API. All API endpoints are prefixed with `/api/v1`.

    ## Authentication
    Protected routes require a Bearer token in the `Authorization` header:
    ```
    Authorization: Bearer <token>
    ```

    Tokens are obtained via `/auth/register` or `/auth/login`.
  version: 0.1.0-alpha.0
  contact:
    name: FinTrak
  license:
    name: AGPL v3.0
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
paths:
...
  /accounts:
    get:
      tags:
        - Accounts
      summary: List all accounts
      description: Get all accounts belonging to the authenticated user
      operationId: getAccounts
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      tags:
        - Accounts
      summary: Create a new account
      description: Create a new bank or credit-card account
      operationId: createAccount
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountRequest"
      responses:
        "201":
          description: Account created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /accounts/{id}:
    put:
      tags:
        - Accounts
      summary: Update an account
      description: Update account details (name, type, bank, currency, color, default status)
      operationId: updateAccount
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: Account ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccountRequest"
      responses:
        "200":
          description: Account updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      tags:
        - Accounts
      summary: Delete an account
      description: Delete an account and its associated payee
      operationId: deleteAccount
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: Account ID
      responses:
        "200":
          description: Account deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /accounts/{id}/export:
    get:
      tags:
        - Accounts
      summary: Export account transactions as CSV
      description: Download all transactions for an account as a CSV file
      operationId: exportAccount
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: Account ID
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /categories/{id}:
    put:
      tags:
        - Categories
      summary: Update a category
      operationId: updateCategory
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Category"
      responses:
        "200":
          description: Category updated
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      tags:
        - Categories
      summary: Delete a category
      description: Uncategorizes its transactions and removes rules pointing at it.
      operationId: deleteCategory
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Category deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ============= Admin (global groups & categories) =============
...
components:
  schemas:
...
    CreateAccountRequest:
      type: object
      required:
        - name
        - accountTypeId
      properties:
        name:
          type: string
        accountTypeId:
          type: string
        bank:
          type: string
        currency:
          type: string
          default: INR
        color:
          type: string
          default: "#06b6d4"
        isDefault:
          type: boolean
        billingDay:
          type: integer
          minimum: 1
          maximum: 31
          default: 1
          description: Credit-card billing day on which each billing cycle ends; only meaningful for credit-card accounts (defaults to 1)
    UpdateAccountRequest:
      type: object
      properties:
        name:
          type: string
        accountTypeId:
          type: string
        bank:
          type: string
        currency:
          type: string
        color:
          type: string
        isDefault:
          type: boolean
        closed:
          type: boolean
          description: Marks the account closed (true) or reopens it (false); omitted keeps the current value
        billingDay:
          type: integer
          minimum: 1
          maximum: 31
          description: Credit-card billing day on which each billing cycle ends; omitted keeps the current value
...
    Account:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        accountTypeId:
          type: string
        accountTypeName:
          type: string
        bank:
          type: string
        currency:
          type: string
        color:
          type: string
        isDefault:
          type: boolean
        closed:
          type: boolean
          description: Marks the account closed (true) or reopens it (false)
        balance:
          type: number
          format: double
        billingDay:
          type: integer
          minimum: 1
          maximum: 31
          description: Credit-card billing day on which each billing cycle ends (default 1)
...
    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        icon:
          type: string
        color:
          type: string
        type:
          type: string
...
    ErrorResponse:
      type: object
      required:
        - errors
      properties:
        errors:
          type: array
          minItems: 1
          items:
            type: object
            required:
              - message
            properties:
              field:
                type: string
              tag:
                type: string
              message:
                type: string
    MessageResponse:
      type: object
      properties:
        message:
          type: string
  responses:
    BadRequest:
      description: Bad request (invalid input)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/register or /auth/login